Payload Target Setting/Completion + Mapping of Targets (#29)
Features added:

Setting of current target
Deleting detected objects
Displaying of completed targets (green highlight)
Outlier highlighting. If some detections are outside of a specified distance of the cluster, they are marked as outliers. There are 3 outlier categories: minor (40m < dist < 65m), mild (65m < dist < 90m), major (dist > 90m). Distances are in meters from the average calculated position of the detections for an object. This is to help combat false positives, which may effect the calculated active target location.
Mapping of detected objects in real time. The map allows for interacting with the objects and visualizing exactly where the detections are. Current target is displayed as the average of the detections for the selected target. Outlier highlighting is displayed here as well.
Real-time display of payload status. Payload panel displays:
- Current target, with its average calculated position for payload drop
- Plane distance and ETA to drop point. Calculated from plane location, airspeed, heading, drop distance (assuming 8m from actual target for now), and active target location